requirements.txt
asgiref==3.6.0
Django==4.1.5
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-widget-tweaks==1.4.12
install==1.3.5
phonenumbers==8.13.5
psycopg2==2.9.5
sqlparse==0.4.3
tzdata==2022.7
